+++ jscomp/runtime/js_undefined.mli
@@ -33,30 +33,29 @@ external return : 'a -> 'a t = "%identity"
 (** Constructs a value of ['a Js.undefined] containing a value of ['a] *)
 
 val testAny : 'a -> bool
-(**
-   Returns [true] if the given value is [empty] ([undefined])
-   @since 1.6.1
-*)
+(** Returns [true] if the given value is [empty] ([undefined]) *)
 
 external empty : 'a t = "#undefined"
 (** The empty value, [undefined] *)
 
 external getUnsafe : 'a t -> 'a = "%identity"
 val getExn : 'a t -> 'a
+val map : f:(('a -> 'b)[@u]) -> 'a t -> 'b t
 
-val bind : 'a t -> (('a -> 'b)[@u]) -> 'b t
-(** Maps the contained value using the given function
+val bind : f:(('a -> 'b t)[@u]) -> 'a t -> 'b t
+(** Bind the contained value using the given function
 
-If ['a Js.undefined] contains a value, that value is unwrapped, mapped to a ['b] using
-the given function [a' -> 'b], then wrapped back up and returned as ['b Js.undefined]
+If ['a Js.undefined] contains a value, that value is unwrapped, mapped to a
+['b] using the given function [a' -> 'b], then wrapped back up and returned as
+['b Js.undefined]
 
 {[
 let maybeGreetWorld (maybeGreeting: string Js.undefined) =
-  Js.Undefined.bind maybeGreeting (fun greeting -> greeting ^ " world!")
+  Js.Undefined.bind maybeGreeting ~f:(fun greeting -> greeting ^ " world!")
 ]}
 *)
 
-val iter : 'a t -> (('a -> unit)[@u]) -> unit
+val iter : f:(('a -> unit)[@u]) -> 'a t -> unit
 (** Iterates over the contained value with the given function
 
 If ['a Js.undefined] contains a value, that value is unwrapped and applied to
@@ -64,7 +63,7 @@ the given function.
 
 {[
 let maybeSay (maybeMessage: string Js.undefined) =
-  Js.Undefined.iter maybeMessage (fun message -> Js.log message)
+  Js.Undefined.iter maybeMessage ~f:(fun message -> Js.log message)
 ]}
 *)