+++ ppx/ast_derive/ast_derive_abstract.ml
@@ -23,22 +23,12 @@
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA. *)
 
 open Import
-
-(* let derivingName = "abstract" *)
-module U = Ast_derive_util
 open Ast_helper
-(* type tdcls = Parsetree.type_declaration list *)
 
-(** For this attributes, its type was wrapped as an option,
-   so we can still reuse existing frame work
-*)
-let get_optional_attrs =
-  [ Ast_attributes.mel_get; Ast_attributes.mel_return_undefined ]
-
-let deprecated_abstract : Parsetree.attribute =
+let deprecated_abstract =
   let loc = Location.none in
   {
-    attr_name = { txt = "alert"; loc = Location.none };
+    attr_name = { txt = "alert"; loc };
     attr_payload =
       PStr
         [
@@ -47,21 +37,13 @@ let deprecated_abstract : Parsetree.attribute =
               "The `@deriving abstract` deriver is deprecated, use `@deriving \
                dynamicKeys` instead."];
         ];
-    attr_loc = Location.none;
+    attr_loc = loc;
   }
 
 let with_deprecation ~is_deprecated attrs =
   match is_deprecated with
   | false -> attrs
   | true -> deprecated_abstract :: attrs
-
-let get_attrs ~is_deprecated =
-  with_deprecation ~is_deprecated
-    Ast_attributes.[ mel_get_arity; unboxable_type_in_prim_decl ]
-
-let set_attrs ~is_deprecated =
-  with_deprecation ~is_deprecated
-    Ast_attributes.[ mel_set; unboxable_type_in_prim_decl ]
 
 let get_pld_type pld_type ~attrs =
   let is_optional = Ast_attributes.has_mel_optional attrs in
@@ -73,118 +55,189 @@ let get_pld_type pld_type ~attrs =
           "`[@mel.optional]' must appear on an option literal type (`_ option')"
   else pld_type
 
-let handleTdcl ~is_deprecated light (tdcl : Parsetree.type_declaration) :
-    Parsetree.value_description list =
-  let loc = tdcl.ptype_loc in
-  let type_name = tdcl.ptype_name.txt in
-  let core_type = U.core_type_of_type_declaration tdcl in
+let derive_js_constructor ?(is_deprecated = false) tdcl =
   match tdcl.ptype_kind with
-  | Ptype_record label_declarations ->
-      let is_private = tdcl.ptype_private = Private in
+  | Ptype_record label_declarations -> (
+      let loc = tdcl.ptype_loc in
       let has_optional_field =
         List.exists
           ~f:(fun (x : Parsetree.label_declaration) ->
             Ast_attributes.has_mel_optional x.pld_attributes)
           label_declarations
       in
-      let setter_accessor, makeType, labels =
+      let makeType, labels =
         List.fold_right
           ~f:(fun
-              ({
-                 pld_name = { txt = label_name; loc = label_loc } as pld_name;
-                 pld_type;
-                 pld_mutable;
-                 pld_attributes;
-                 pld_loc;
-               } :
-                Parsetree.label_declaration)
-              (acc, maker, labels)
+              {
+                pld_name = { txt = label_name; loc = _ } as pld_name;
+                pld_type;
+                pld_attributes;
+                pld_loc;
+                _;
+              }
+              (maker, labels)
             ->
-            let prim_as_name, newLabel =
+            let newLabel =
               match
                 Ast_attributes.iter_process_mel_string_as pld_attributes
               with
-              | None -> (label_name, pld_name)
-              | Some new_name -> (new_name, { pld_name with txt = new_name })
+              | None -> pld_name
+              | Some new_name -> { pld_name with txt = new_name }
+            in
+            let is_optional = Ast_attributes.has_mel_optional pld_attributes in
+            let maker =
+              if is_optional then
+                let pld_type = get_pld_type ~attrs:pld_attributes pld_type in
+                Typ.arrow ~loc:pld_loc (Optional label_name) pld_type maker
+              else Typ.arrow ~loc:pld_loc (Labelled label_name) pld_type maker
+            in
+            (maker, (is_optional, newLabel) :: labels))
+          label_declarations
+          ~init:
+            ( (let core_type =
+                 Ast_derive_util.core_type_of_type_declaration tdcl
+               in
+               if has_optional_field then [%type: unit -> [%t core_type]]
+               else core_type),
+              [] )
+      in
+      match tdcl.ptype_private with
+      | Private -> []
+      | Public ->
+          let myPrims =
+            Ast_external_mk.pval_prim_of_option_labels labels has_optional_field
+          in
+          [
+            Val.mk ~loc
+              { loc; txt = tdcl.ptype_name.txt }
+              ~attrs:
+                (with_deprecation ~is_deprecated
+                   [ Ast_attributes.unboxable_type_in_prim_decl ])
+              ~prim:myPrims makeType;
+          ])
+  | Ptype_abstract | Ptype_variant _ | Ptype_open ->
+      (* Looks obvious that it does not make sense to warn *)
+      []
+
+let derive_getters_setters =
+  let get_optional_attrs =
+    (* For these attributes, its type was wrapped as an option,
+       so we can still reuse existing framework *)
+    [ Ast_attributes.mel_get; Ast_attributes.mel_return_undefined ]
+  in
+  let get_attrs =
+    Ast_attributes.[ mel_get_arity; unboxable_type_in_prim_decl ]
+  in
+  let set_attrs = Ast_attributes.[ mel_set; unboxable_type_in_prim_decl ] in
+  fun ?(is_deprecated = false) ~light tdcl ->
+    match tdcl.ptype_kind with
+    | Ptype_record label_declarations ->
+        let loc = tdcl.ptype_loc in
+        let core_type = Ast_derive_util.core_type_of_type_declaration tdcl in
+        List.fold_right
+          ~f:(fun
+              {
+                pld_name = { txt = label_name; loc = label_loc } as pld_name;
+                pld_type;
+                pld_mutable;
+                pld_attributes;
+                pld_loc;
+              }
+              acc
+            ->
+            let prim_as_name =
+              match
+                Ast_attributes.iter_process_mel_string_as pld_attributes
+              with
+              | None -> label_name
+              | Some new_name -> new_name
             in
             let prim = [ prim_as_name ] in
-            let is_optional = Ast_attributes.has_mel_optional pld_attributes in
-
-            let maker, acc =
-              if is_optional then
-                let optional_type = pld_type in
-                let pld_type = get_pld_type ~attrs:pld_attributes pld_type in
-                ( Typ.arrow ~loc:pld_loc (Optional label_name) pld_type maker,
-                  Val.mk ~loc:pld_loc
-                    (if light then pld_name
-                     else { pld_name with txt = pld_name.txt ^ "Get" })
-                    ~attrs:get_optional_attrs ~prim
-                    [%type: [%t core_type] -> [%t optional_type]]
-                  :: acc )
-              else
-                ( Typ.arrow ~loc:pld_loc (Labelled label_name) pld_type maker,
-                  Val.mk ~loc:pld_loc
-                    (if light then pld_name
-                     else { pld_name with txt = pld_name.txt ^ "Get" })
-                    ~attrs:(get_attrs ~is_deprecated)
-                    ~prim:
-                      ((* Not needed actually*)
-                       Melange_ffi.External_ffi_types.ffi_mel_as_prims
-                         [ Melange_ffi.External_arg_spec.dummy ]
-                         Return_identity
-                         (Js_get
-                            { js_get_name = prim_as_name; js_get_scopes = [] }))
-                    [%type: [%t core_type] -> [%t pld_type]]
-                  :: acc )
-            in
-            let is_current_field_mutable = pld_mutable = Mutable in
             let acc =
-              if is_current_field_mutable then
+              if Ast_attributes.has_mel_optional pld_attributes then
+                let optional_type = pld_type in
+                Val.mk ~loc:pld_loc
+                  (if light then pld_name
+                   else { pld_name with txt = pld_name.txt ^ "Get" })
+                  ~attrs:get_optional_attrs ~prim
+                  [%type: [%t core_type] -> [%t optional_type]]
+                :: acc
+              else
+                Val.mk ~loc:pld_loc
+                  (if light then pld_name
+                   else { pld_name with txt = pld_name.txt ^ "Get" })
+                  ~attrs:(with_deprecation ~is_deprecated get_attrs)
+                  ~prim:
+                    ((* Not needed actually*)
+                     Melange_ffi.External_ffi_types.ffi_mel_as_prims
+                       [ Melange_ffi.External_arg_spec.dummy ]
+                       Return_identity
+                       (Js_get
+                          { js_get_name = prim_as_name; js_get_scopes = [] }))
+                  [%type: [%t core_type] -> [%t pld_type]]
+                :: acc
+            in
+            match pld_mutable with
+            | Mutable ->
                 let pld_type = get_pld_type pld_type ~attrs:pld_attributes in
                 let setter_type =
                   [%type: [%t core_type] -> [%t pld_type] -> unit]
                 in
                 Val.mk ~loc:pld_loc
                   { loc = label_loc; txt = label_name ^ "Set" } (* setter *)
-                  ~attrs:(set_attrs ~is_deprecated) ~prim setter_type
+                  ~attrs:(with_deprecation ~is_deprecated set_attrs)
+                  ~prim setter_type
                 :: acc
-              else acc
-            in
-            (acc, maker, (is_optional, newLabel) :: labels))
-          label_declarations
-          ~init:
-            ( [],
-              (if has_optional_field then [%type: unit -> [%t core_type]]
-               else core_type),
-              [] )
-      in
-      if is_private then setter_accessor
-      else
-        let myPrims =
-          Ast_external_mk.pval_prim_of_option_labels labels has_optional_field
-        in
-        let myMaker =
-          Val.mk ~loc { loc; txt = type_name }
-            ~attrs:
-              (with_deprecation ~is_deprecated
-                 [ Ast_attributes.unboxable_type_in_prim_decl ])
-            ~prim:myPrims makeType
-        in
-        myMaker :: setter_accessor
-  | Ptype_abstract | Ptype_variant _ | Ptype_open ->
-      (* Looks obvious that it does not make sense to warn *)
-      []
+            | Immutable -> acc)
+          label_declarations ~init:[]
+    | Ptype_abstract | Ptype_variant _ | Ptype_open ->
+        (* Looks obvious that it does not make sense to warn *)
+        []
 
-let handleTdclsInStr ~is_deprecated ~light _rf tdcls =
+let derive_js_constructor_str _rf tdcls =
   List.fold_right
     ~f:(fun tdcl sts ->
-      match handleTdcl ~is_deprecated light tdcl with
-      | value_descriptions -> List.map ~f:Str.primitive value_descriptions @ sts)
+      let value_descriptions = derive_js_constructor tdcl in
+      List.map ~f:Str.primitive value_descriptions @ sts)
     tdcls ~init:[]
 
-let handleTdclsInSig ~is_deprecated ~light _rf tdcls =
+let derive_js_constructor_sig _rf tdcls =
   List.fold_right
     ~f:(fun tdcl sts ->
-      match handleTdcl ~is_deprecated light tdcl with
-      | value_descriptions -> List.map ~f:Sig.value value_descriptions @ sts)
+      let value_descriptions = derive_js_constructor tdcl in
+      List.map ~f:Sig.value value_descriptions @ sts)
+    tdcls ~init:[]
+
+let derive_getters_setters_str ~light _rf tdcls =
+  List.fold_right
+    ~f:(fun tdcl sts ->
+      let value_descriptions = derive_getters_setters tdcl ~light in
+      List.map ~f:Str.primitive value_descriptions @ sts)
+    tdcls ~init:[]
+
+let derive_getters_setters_sig ~light _rf tdcls =
+  List.fold_right
+    ~f:(fun tdcl sts ->
+      let value_descriptions = derive_getters_setters ~light tdcl in
+      List.map ~f:Sig.value value_descriptions @ sts)
+    tdcls ~init:[]
+
+let derive_abstract_str ~light _rf tdcls =
+  List.fold_right
+    ~f:(fun tdcl sts ->
+      let cstr_descriptions = derive_js_constructor ~is_deprecated:true tdcl in
+      let value_descriptions =
+        derive_getters_setters ~is_deprecated:true ~light tdcl
+      in
+      List.map ~f:Str.primitive (cstr_descriptions @ value_descriptions) @ sts)
+    tdcls ~init:[]
+
+let derive_abstract_sig ~light _rf tdcls =
+  List.fold_right
+    ~f:(fun tdcl sts ->
+      let cstr_descriptions = derive_js_constructor ~is_deprecated:true tdcl in
+      let value_descriptions =
+        derive_getters_setters ~is_deprecated:true ~light tdcl
+      in
+      List.map ~f:Sig.value (cstr_descriptions @ value_descriptions) @ sts)
     tdcls ~init:[]