+++ ppx/ast_derive/ast_derive_abstract.mli
@@ -24,15 +24,31 @@
 
 open Ppxlib
 
-val handleTdclsInStr :
-  is_deprecated:bool ->
+val derive_abstract_str :
   light:bool ->
   Asttypes.rec_flag ->
   Parsetree.type_declaration list ->
   Parsetree.structure
 
-val handleTdclsInSig :
-  is_deprecated:bool ->
+val derive_abstract_sig :
+  light:bool ->
+  Asttypes.rec_flag ->
+  Parsetree.type_declaration list ->
+  Parsetree.signature
+
+val derive_js_constructor_str :
+  Asttypes.rec_flag -> Parsetree.type_declaration list -> Parsetree.structure
+
+val derive_js_constructor_sig :
+  Asttypes.rec_flag -> Parsetree.type_declaration list -> Parsetree.signature
+
+val derive_getters_setters_str :
+  light:bool ->
+  Asttypes.rec_flag ->
+  Parsetree.type_declaration list ->
+  Parsetree.structure
+
+val derive_getters_setters_sig :
   light:bool ->
   Asttypes.rec_flag ->
   Parsetree.type_declaration list ->